+++ exonum/src/events/network.rs
@@ -22,6 +22,7 @@ use std::{cell::RefCell, collections::HashMap, io, net::SocketAddr, rc::Rc, time
 
 use super::{error::{into_other, log_error, other_error, result_ok},
             to_box};
+use crypto::PublicKey;
 use events::noise::{Handshake, HandshakeParams, NoiseHandshake};
 use helpers::Milliseconds;
 use messages::{Any, Connect, Message, RawMessage};
@@ -38,7 +39,7 @@ pub enum NetworkEvent {
 
 #[derive(Debug, Clone)]
 pub enum NetworkRequest {
-    SendMessage(SocketAddr, RawMessage),
+    SendMessage(SocketAddr, RawMessage, PublicKey),
     DisconnectWithPeer(SocketAddr),
     Shutdown,
 }
@@ -147,8 +148,7 @@ impl ConnectionsPool {
                 Ok(sock)
             })
             .and_then(move |sock| {
-                let handshake = NoiseHandshake::initiator(&handshake_params);
-                handshake.send(sock)
+                NoiseHandshake::initiator(&handshake_params).send(sock)
             })
             // Connect socket with the outgoing channel
             .and_then(move |stream| {
@@ -256,13 +256,16 @@ impl RequestHandler {
         handshake_params: &HandshakeParams,
     ) -> RequestHandler {
         let mut cancel_sender = Some(cancel_sender);
-        let handshake_params = handshake_params.clone();
         let outgoing_connections = ConnectionsPool::new();
+        let handshake_params = handshake_params.clone();
         let requests_handler = receiver
             .map_err(|_| other_error("no network requests"))
             .for_each(move |request| {
                 match request {
-                    NetworkRequest::SendMessage(peer, msg) => {
+                    NetworkRequest::SendMessage(peer, msg, remote_key) => {
+                        let mut handshake_params = handshake_params.clone();
+                        handshake_params.set_remote_key(remote_key);
+
                         let conn_tx = outgoing_connections
                             .get(peer)
                             .map(|conn_tx| conn_fut(Ok(conn_tx).into_future()))