+++ exonum/src/lib.rs
@@ -35,8 +35,6 @@ extern crate clap;
 extern crate colored;
 extern crate cookie;
 extern crate env_logger;
-#[macro_use]
-extern crate exonum_profiler;
 extern crate exonum_rocksdb as rocksdb;
 extern crate exonum_sodiumoxide as sodiumoxide;
 #[macro_use]